(SERVER-657) Bump lein-ezbake dependency to 0.2.10
This commit bumps the lein-ezbake dependency to version 0.2.10.  This is
needed in order to allow Puppet Server packaging to include an updated
version of the foreground command which allows the HOME environment
variable to be set for the puppet user when the foreground subcommand is
executed.  Without this change, attempts to execute the foreground
subcommand on Ubuntu could result in an inappropriate value for the
HOME environment variable being used.