(MAINT) Add install options to 'install_dev_repos'
This commit configures options explicitly for the
'30_install_dev_repos.rb' for package installs.  This was done
specifically to configure the use of the 'PC1' repo since this is
required for use with AIO testing.  The need for this change was
introduced by the beaker refactoring done for BKR-184 which removed PC1
from the default list of repos that beaker checks during installation.