(MAINT) Add install options to 'install_dev_repos' for Puppet 3.x
This commit configures the PC1 repo for the Puppet 3.x compatibility
pre-suite as the previous commit did for the full Puppet Server
acceptance pre-suite.  This commit also changes the options definition
to merge the global defaults in with just the one repos setting which
needs to be overridden for use with the PC1 repo.