(SERVER-297) Consolidate JRuby environment handling
Without this patch the manner in which environment variables are handled
in the JRuby portions of puppet-server vary across the use cases of the
`puppetserver {gem,irb,ruby}` command line utilities in addition to the
service itself.  This is a problem because the behaviors have not kept
in sync with one another over time so there has been divergence at
times, e.g. `puppetserver gem env` currently fails because PATH is unset
in that scenario but set in others, e.g. the production service.

This patch addresses the problem by consolidating all of the environment
handling behavior into a single authoritative method.  The behavior
itself remains unchanged, GEM_HOME is set based on the config file
value, PATH and HOME are the only variables allowed to flow through to
JRuby unmodified, all other variables are filtered out, and JARS_REQUIRE
and JARS_NO_REQUIRE are set appropriately.

In addition to the environment handling, this patch also consolidates
the JRuby load path handling in a similar fashion across the scenarios
of command line utilities and the production service.