(SERVER-721) Consolidate JRuby environment handling
Without this patch the manner in which environment variables are handled
in the JRuby portions of puppet-server vary across the use cases of the
`puppetserver {gem,irb,ruby}` command line utilities in addition to the
service itself.  This is a problem because the behaviors have not kept
in sync with one another over time so there has been divergence at
times, e.g. `puppetserver gem env` currently fails because PATH is unset
in that scenario but set in others, e.g. the production service.

This patch addresses the problem by consolidating all of the environment
handling behavior into a single authoritative method.  The behavior
itself remains unchanged, GEM_HOME is set based on the config file
value, PATH and HOME are the only variables allowed to flow through to
JRuby unmodified, all other variables are filtered out, and JARS_REQUIRE
and JARS_NO_REQUIRE are set appropriately.

In addition to the environment handling, this patch also consolidates
the JRuby load path handling in a similar fashion across the scenarios
of command line utilities and the production service.

(maint) Fix managed-load-path schema errors

Fix this error:

    Caused by: clojure.lang.ExceptionInfo: Output of managed-load-path does not match schema: [(not (instance? java.lang.String a-java.io.File)) (not (instance? java.lang.String a-java.io.File)) (not (instance? java.lang.String a-java.io.File)) (not (instance? java.lang.String a-java.io.File)) (not (instance? java.lang.String a-java.io.File)) (not (instance? java.lang.String a-java.io.File))]
     at puppetlabs.services.jruby.jruby_puppet_internal$eval9067$managed_load_path__9068.invoke (jruby_puppet_internal.clj:69)
        puppetlabs.services.jruby.jruby_puppet_internal$prep_scripting_container.invoke (jruby_puppet_internal.clj:89)
        puppetlabs.services.jruby.jruby_puppet_internal$empty_scripting_container.invoke (jruby_puppet_internal.clj:101)
        puppetlabs.services.jruby.jruby_puppet_internal$create_scripting_container.invoke (jruby_puppet_internal.clj:117)
        puppetlabs.services.jruby.jruby_puppet_internal$eval9122$create_pool_instance_BANG___9123$fn__9124.invoke (jruby_puppet_internal.clj:156)
        puppetlabs.services.jruby.jruby_puppet_internal$eval9122$create_pool_instance_BANG___9123.invoke (jruby_puppet_internal.clj:141)

(maint) Remove unused definitions from jruby-puppet-core

These defs moved over to jruby-puppet-internal in
23154f42f24351e55a329cf52663d291c4c5312f and are no longer used from
jruby-puppet-core.

(SERVER-297) Fix up LOAD_PATH behavior

Without this patch the load path behavior is incorrect because the use
of normalized and absolute paths is incompatible with the use of loading
resources from a sealed jar.  This patch addresses the problem by
restoring the previous behavior of implicitly relying in JRuby's
LoadService implementation.

For more information, particularly on the combination and interaction of
load path, class path, and ruby require, please see:

https://github.com/jruby/jruby/blob/1.7.20/core/src/main/java/org/jruby/runtime/load/LoadService.java

(SERVER-297) Add jruby subcommand integration tests

Without this patch there are not tests covering the behavior of the irb,
ruby, and gem JRuby subcommands.  This patch addresses the problem by
adding integration tests that assert against standard output and the
exit status.

NB this patch differs from the patch against master in that the
ruby-load-path setting is under the os-settings configuration section in
this patch while it is under the jruby-puppet section in the patch
against master for puppetserver 2.0

(SERVER-297) Check if subcommand exists before loading

Without this patch the subcommand string given to the cli-run! command
may not actually exist in the assumed directory of
META-INF/jruby.home/bin/  This patch checks before trying to execute the
given subcommand.

(maint) Mark j_p_internal functions public / private

Addressing PR comments, sorted by using "find usages" in cursive.

(maint) Add compile-mode, compat-version defs

Without this patch the initialization logic for the JRuby instances in
new-main and prep-scripting-container are almost identical.  This is a
problem because they should be identical and the compatibility version
and compile mode are not referenced in a way that changing one will also
change the other.

This patch addresses the problem by making the behavior identical
without going so far as to providing a single function that operates on
both classes of objects.

(maint) use schema/maybe instead of schema/pred nil?

(SERVER-297) Add acceptance tests for env behavior

Without this patch there aren't any acceptance tests that validate the
behavior of environment handling.  This patch addresses the problem by
adding a test for the irb and ruby commands.  The gem command is already
tested via the `gem env` command and is difficult to inspect the
environment from since it doesn't evaluate arbitrary code like irb and
ruby can.

(maint) Fix failing FOO_DEBUG irb and ruby CLI tests

The smoke tests are failing checking that FOO_DEBUG is being preserved
for the irb and ruby subcommands.  This is a problem because the
behavior has been consolidated around the behavior of the production
scripting container which is to filter out variables like FOO_DEBUG.
This patch addresses the problem by changing the test to assert the
desired behavior.

(maint) Fix failing smoke test on undefined var output

Fixes:

    #<NameError: undefined local variable or method `output'

(maint) Fix failing smoke test on irb -e

Fixes this error:

    IRB::UnrecognizedSwitch: Unrecognized switch: -e

irb reads from STDIN, it doesn't have a -e flag like the ruby command
does.

(maint) Add jruby-schemas/EnvMap