(SERVER-297) Fix up LOAD_PATH behavior
Without this patch the load path behavior is incorrect because the use
of normalized and absolute paths is incompatible with the use of loading
resources from a sealed jar.  This patch addresses the problem by
restoring the previous behavior of implicitly relying in JRuby's
LoadService implementation.

For more information, particularly on the combination and interaction of
load path, class path, and ruby require, please see:

https://github.com/jruby/jruby/blob/1.7.20/core/src/main/java/org/jruby/runtime/load/LoadService.java