(maint) Minor improvements to File Sync Client
This commit contains minor improvements to the File Sync Client.  Most
importantly, the code which initializes the client's repository now
tolerates an existing empty directory.  This may not be needed in
production, but there's no reason for this restriction - we can clone
into an empty directory just fine, and it makes testing more difficult.

This commit also contains a couple minor logging improvements.