(maint) Prevent callback registration after start
Prevent services from registering a callback with the File Sync Client
after it has already been started.  Previously, the call to
register-callback! would work just fine, but it would have no effect -
that is, the callback would not ever be invoked, since the set of
callbacks is effectively frozen once the service is started.  This
additional error handling simply makes that fact more explicit.

These changes also introduce egregious Wolfe Principle violations; the
underlying cause of these is TK-226.  In this case, the File Sync
Client's scheduled job is never being stopped, and this causes some
errors to be logged.