(SERVER-297) Remove duplicate entries at the head of $LOAD_PATH
Without this patch there are duplicate entries for the system ruby
library paths at the head of the $LOAD_PATH.  This behavior exists to
prevent accidentally loading libraries from the OS MRI copy of Ruby and
was necessary back when puppetserver loaded puppet code from the OS ruby
library paths.

This is no longer necessary as puppet code is isolated in PC1 and exists
outside of any Ruby library paths.  This patch addresses the problem by
removing the duplicate entries from the head of the LOAD_PATH restoring
the default behavior of JRuby.