(PE-8993) Pass in server-repo-url to storage core functions
Previously, we were passing the `server-url` (the scheme, hostname, and port
for the storage service) to the storage core functions to be used when adding
a submodule. However, we then combined it with a hardcoded path for the git
servlet. This commit changes this to construct a `server-repo-url` out of the
`server-url` and `repo-path-prefix`, and passes this in, so that we no longer
use a hardcoded path to the git servlet in the submodule add.