(maint) Add compile-mode, compat-version defs
Without this patch the initialization logic for the JRuby instances in
new-main and prep-scripting-container are almost identical.  This is a
problem because they should be identical and the compatibility version
and compile mode are not referenced in a way that changing one will also
change the other.

This patch addresses the problem by making the behavior identical
without going so far as to providing a single function that operates on
both classes of objects.