(PE-8993) Return SHA for commit of submodule within parent repo
Previously, when doing a publish on a repo with submodules, we reutrned the
SHA for the submodule's own git repository after doing an `add-and-commit` on
it. However, if multiple "publish" requests came in at the sae time, there's a
chance that this SHA might not actually reflect the SHA of the submodule that
actually got updated on the parent repo. This commit updates the SHAs returned
to instead be the SHA of the new HEAD of the submodule within the parent repo.