(SERVER-718) Load facter jar into system classloader classpath
This commit adds some logic which searches for a file named 'facter.jar'
in the ruby load path and, if found, adds the file to the system
classloader's classpath.  This is needed for compatibility with Facter
3+.  The jar needs to be in the system classloader's classpath in order
for Facter's backing native shared library to only be loaded once by a
classloader, avoiding an error which would otherwise occur if attempts
were made to load the library into more than 1 JRuby container instance
classloader.