(maint) Remove race condition in server-side corruption integration test
In the server-side corruption integration test, there is a possibility that a
sync can happen between when we push to the storage repo and when the
"corrupted" repo's storage-side bare repo is renamed. This changes the test to
check the agent's state, rather than the status of the client repo, ensuring
that the test will still pass if this sync happens.