(SERVER-768) trapperkeeper-authorization POC
This patch replaces the client whitelist authorization strategy with a
rules based strategy using Brice's [trapperkeeper-authorization
library][tk-authz].  This patch is not meant to be the final solution
but is instead intended to evaluate and scope additional work on the
authorization behavior.

This patch replaces the whitelist checking ring middleware with a
middleware that loads rules from /etc/puppetlabs/puppetserver/rules.conf
if present.  If the file does not exist then hard-coded default rules
which deny everything are used.

Hacks for the POC related to the complexity of the changes required:

This patch removes the previous behavior of being able to turn off
access control entirely.  Similar behavior is provided using permissive
authorization rules, however this patch requires an SSL certificate from
the client.  If no client certificate is present then the request is
forbidden.

 * Decision needed: Do we need to preserve the behavior of disabling authz
   entirely?

The following tests are removed because they make assumptions about
client-whitelist.  These tests are expected to be fairly straightforward
to update.  The test set-up behavior needs to replace the
client-whitelist configuration with an approach that configures the
authz rules.  All of these tests should be able to be treated in the
same manner:

 * puppet-environments-int-test
 * puppet-admin-int-test
 * jruby-pool-int-test

There will be some complexity around this test, which may require
special handling relative to the mechanical nature of replacing the
client-whitelist configuration:

 * jruby-pool-int-test (The wait-for-new-pool function enters an
   infinite loop with the certificate whitelist behavior disabled.)

Similarly, the legacy-routes-test/legacy-routes test completely disables
access control, however there should be no additional complexity fixing
this test beyond either adopting the new way to disable AuthZ entirely
(if we do that) or configuring the test to make an authorized request by
adjusting the rules.

 * legacy-routes-test/legacy-routes

[tk-authz]: https://github.com/masterzen/trapperkeeper-authorization