(PE-10499) Storage Service status: add client info
Add information about clients to the Storage Service's /status response.
The implementation works by simply transmitting a bit of the Agent's
state (on the client) to the server during the /latest-commits request.
To accommodate this change, the verb of the request is changed from GET to POST.

These changes also required updating the AgentState schema to more
accurately correspond to the states the Agent can be in.