(PE-10610) Make file sync service work with nested .git dirs
Add an integration test that tests that we can manage directories with nested
.git directories in both a parent repo and in submodules.

Update file sync to use the NoGitLinksWorkingTreeIterator to make this test
(and all the others) pass. There are a couple quirky things that had to be
done to make this happen:

* By using the NoGitlinksWorkingTreeIterator we can no longer use `.setAll true`
when doing a commit. `.setAll true` will add all files that have been modified
to the index, and it will also remove from the index all files that have been
deleted in the working tree. This means that we must instead iterate through
all files removed from the working tree and do a `git rm` on each, since `git
rm` doesn't support `.`.
* We can't use the NoGitlinksWorkingTreeIterator when doing a `git add` on the
submodules-dir, otherwise these would not be added as submodules. Instead we
have to "reset" this directory, then re-add it with the default
FileTreeIterator.