(SERVER-768) Improve debug logging of the authz behavior
Without this patch the authorization behavior isn't logged very well at
all.  The size of the request map in the logs obfuscates valuable
information and the message coming back from the
trapperkeeper-authorization library is discarded.

This patch implements utility methods that formats the request itself
and the results of the authorization check in a manner suitable for
human readability and for processing as JSON.