(maint) Add with-test-logging to remove error message in test
Whenever we start up the client service with the storage service a "Connection
Refused" error message is logged, because the client service does not have a
dependency on the storage service, and it is faster to start on the storage
service. This commit adds a `with-test-logging` so that this error is not
logged when tests are run.