(PE-8226) Add `action` argument to borrow fns
This commit adds a new argument, `action`, to the functions relating
to borrowing a JRuby instance.  The value will be used for logging,
metrics, debugging, and similar purposes; so that at any point in
time we can ask the system how many JRuby instances are loaned out,
and what they are being used for.