(PE-8226) Add `action` argument to JRuby borrow/return fns
This commit adds a new argument, `action`, to the functions relating
to borrowing and returning a JRuby instance.  The value will be used
for logging, metrics, debugging, and similar purposes; so that at any
point in time we can ask the system how many JRuby instances are loaned out,
and what they are being used for.