(SERVER-815) Add getClassInfo methods to JRubyPuppet interface
This commit adds two new methods to the JRubyPuppet interface -
`getClassInfoForEnvironment` and `getClassInfoForAllEnvironments` - to
the JRubyPuppet interface.  A corresponding implementation of those in
the `Puppet::Server::Master` Ruby class uses the
`Puppet::InfoService::ClassInformationService` to get class information
for available manifests in environments.

This commit also moves the ruby/puppet submodule forward to a newer
commit in the Ruby Puppet repo which contains the initial implementation
of the `Puppet::InfoService::ClassInformationService`.