(PE-11231) Deal with race conditions in test
The server-side-corruption-test has race conditions due to the
asynchronous nature of the Client Service's periodic sync job and the
pattern employed by the test of adding watches to examine the agent's
state.  This commit changes the test to, instead, poll against the
Client Service's status endpoint until a response is received which
matches the state expected by the test.  While this paradigm shift in
the test does not eliminate race conditions entirely (i.e., the polling
could never return an expected response), it drastically improves the
situation: previously, subtle timing issues could could result in a lost
race which would cause the test to fail, even though the service was
performing in a completely reasonable and expected way.  Now, the test
will only fail due to a lost race if an unreasonable amount of time
elapses while the test is waiting (and polling) for the system to reach
its expected state.