(SERVER-815) Bump JVM max heap size in project.clj to 2g
This commit bumps the maximum JVM heap size to 2g in the project.clj.
This would allow `lein test` runs to use up to 2g rather than the 1g
that they are limited today.  This is done in hopes of counteracting
some of the 'Out of Memory' JVM errors that we've been seeing when
running unit/integration tests via Travis CI.