(maint) Update acceptance helper to use puppet-agent version from nightlies
Previously, puppet-server stable tests were running against a build of
puppet-agent that was from the per-commit pipeline. This meant that it was not
available on all platforms we run puppet-server stable acceptance tests against.
This commit bumps the puppet build version to a puppet-agent package version in
nightlies, which is built for all the platforms.