(SERVER-835) Integration tk-auth configuration settings
This commit integrates Puppet Server with trapperkeeper-authorization
for authorizing access to both Clojure-level and core Ruby Puppet-hosted
endpoints.  By default, the client-whitelists are still honored, if
present, for puppet-admin and certificate-status(es) Clojure endpoints.
A new optional setting, use-legacy-auth-conf, can be set to 'false' in
order to have tk-auth used for authorizing core Ruby Puppet-hosted
endpoints.  The default for the use-legacy-auth-conf setting, however,
is 'true'.