(SERVER-835) Updated get-handler to take handler to wrap as argument
This commit changes the work done in previous commits to add a
`get-handler` to the master and CA core namespaces.  The previous work
encapsulated the creation of the routes and handler.  This approach
won't work for the PE master implementation, however, since it needs to
construct the routes itself so that it can wrap metrics middleware
around them.  The work in this commit refactors `get-handler` into
`get-wrapped-handler` whose responsibility now is just to wrap common
middleware - tk-authorization, puppet version, etc. - around a base
handler with encapsulated routes which is passed in as an argument.