(SERVER-835) Only put certificate-status in access-control map
This commit changes the `config->ca-settings` function to only associate
the contents of the `:certificate-status` key with the `access-control`
map instead of just the entire contents of the `certificate-authority`
section.  Associating the entire `certificate-authority` section was a
problem with PE Puppet Server since its `reverse-proxy-ca-service`
implementation of the CA service protocol has other proxy-related
content that should not be stored in the `access-control` key.