(SERVER-860) Introduce ReentrantReadWriteLock into JRubyPool data structure
Move the locking logic relating to atomic code deploys into an encapsulated
data structure that manages the JRuby pool.

Add a LockablePool interface that lays out the API this new data structure
needs to fulfill all our requirements.

Change JRubyPool to implement this interface. The current implementation uses
a LinkedBlockingDeque (which the previous implementation also used) as well as
a ReentrantReadWriteLock. By including the ReentrantReadWriteLock in this data
structure, we can remove the trapperkeeper locking service we had added to
pe-puppetserver; pe-puppetserver will instead use the `lock()` and `unlock()`
methods provided by this data structure to perform write locks.

The `borrowItem()/borrowItemWithTimeout()` methods of this implementation do a
`takeFirst` or `pollFirst` from the underlying deque and also lock a read
lock. `returnItem()` does `putFirst` onto the deque and unlocks a read lock.
`insertPill()` does a `putFirst` without locking a read lock; to be used when
we insert Poison Pills when flushing the pool (when a read lock is not held).
These methods are all very similar to the methods we were using previously,
but with the addition of manipulating the lock.