(SERVER-860) Comment out race in pool-lock-reentrant-with-many-borrows-test
Much of this test was already commented out because the current implementation
does not support reentrant borrows. This comments out the race condition in
the test; we'll uncomment this and worry about race conditions when we move to
an implementation that is reentrant.

Also clarify comment on JRubyPool's lock() method.