(SERVER-860) Reorder operations in `register()`
There is a race condition in the implementation of `register()` between adding
an instance to the set of registered elements and adding an instance to the
pool. We want adding the instance to the set of registered elements to happen
first, because we want to ensure that all instances added to the pool have
been registered so that they are always flushed when the environment cache
flush is requested.