(PE-11857) Add with-lock macro to JRuby Service
Add `with-lock` macro to JRuby service, which can be used by other services to
acquiring a lock on the JRubyPool, execute the body, then release the lock.

Add tests that `with-lock` actually does lock the write lock from the
JRubyPool, and add basic tests that `borrow-instance` and `return-instance`
affect the lock in the ways we expect. Add a `get-lock` helper function to get
access to the lock for tests.