(SERVER-914) Allow middleware params in jruby requests
Prior to this commit, the JRuby request handler code would
filter out any :params that had been added to the ring request
map by other middleware.  We now have concrete use cases for
this capability, so this commit loosens the restrictions and
allows other middlewares to add to :params, and have those
params still be visible to JRuby.