(SERVER-964) Support multi-webserver config
Previously, Puppet Server explicitly checked the contents of the
`:webserver` key in determining whether or not it needed to
override webserver settings with those from ruby puppet, meaning
that the settings would always be overridden if a multi-webserver
style configuration is used.

This commit modifies Puppet Server to use the contents of the
configuration at `[:webserver :puppet-server]`, or `:webserver` if
`[:webserver :puppet-server]` does not exist so as to support
multi-webserver configs.