(SERVER-989) Move JRuby locking logic into `core`
This commit moves the logic for locking/unlocking the JRuby pool
out of the macro and into the `core` namespace.  There are no
functional changes involved here, but in a subsequent commit we'll
be adding some event notification code to the locking/unlocking,
and it would have started to get a little unwieldy to do all of
that from inside of the macro.