(PDB-2071) Add PuppetDBTerminusState class
In order to support high availability, the PuppetDB terminus needs
to implement sticky failover. That is, after failing over to another
PuppetDB instance, it should keep using that instance for future
requests as long as possible. This is tricky when running in
PuppetServer, since it hosts JRuby instances. 

This adds the PuppetDBTerminusState java class as a place for the
PuppetDB terminus to put the data it needs to share between all
JRubies in the process.