(MAINT) Remove 'finish' assertion from the foreground acceptance test
This commit removes an assertion that the "Finished shutdown sequence"
message appears in the output of the puppetserver foreground test.  This
message has been seen to sometimes not appear when puppetserver is shut
down - because an OS may kill the process early if puppetserver's
shutdown hook takes too long to run.  For the OSes where this is
problematic - apparently RedHat-based ones but maybe not Debian-based
ones? - there doesn't appear to be any way to prolong the wait time
when running via "timeout" or pressing CTRL-C in the foreground.