(MAINT) Upgrade lein-ezbake
Utopic has reached EOL and we no longer build packages for it, yet our
version of ezbake was old enough that it still attempted to build
packages for it and would fail.

This commit bumps us up to a newer version of lein-ezbake which will
allow for local ezbake builds to work properly. Jenkins CI isn't failing
because we override the platforms in ci-job-configs and thus don't rely
on the default platforms defined in lein-ezbake.