(SERVER-1006) Unregister jruby instances on flush
Prior to this commit, when a JRuby instance was added to the pool,
it was also added to the list of "registered" instances for the pool
so that we could find an authoritative list of all of the instances
even when some were out on loan.

However, we did not have the corresponding logic to *unregister*
instances from this list when an instance was flushed via
max-requests-per-instance.  This resulted in a reference being
kept to the flushed instances even after they were flushed,
meaning that we could not reclaim all of the memory during
garbage collection.  This caused memory to leak over time.

This commit adds code to unregister instances when they are
flushed, so that we do not retain the extra reference to the
flushed instances, and we are able to garbage collect them
and free up the memory.