(MAINT) Introduce PuppetDB testing
This commit adds support for installing PuppetDB and writing tests.
Since PuppetDB packages aren't built for as many platforms as Puppet
Server we need to keep a list of supported platforms to protect the
pre-suite step and any tests against PuppetDB from breaking CI.

The PuppetDB in this commit is a relatively simple smoke test that
validates Puppet Server can communicate with PuppetDB and send it agent
information.