(SERVER-1016) Add pool releaseItem variant for unlocking item w/o return
This commit renames the returnItem method in the JRubyPool and
LockablePool to releaseItem and adds a variant of the method which can
be used to determine whether the item should be released back into the
pool or just discarded.  For cases where an item is flushed, the item is
just discarded without going back into the pool - but with the
associated release lock for the thread being unlocked.