(SERVER-862) Added call to signal notEmpty when write lock unlocked
This commit adds a call to signal the 'notEmpty' condition for all
waiters when the write lock is unlocked.  This is needed to compensate
for cases where 'notEmpty' signals received while the write lock is held
are received but essentially ignored, leading other borrowers to not be
notified by a 'notEmpty' signal when they should be.

In previous commits, if a 'notEmpty' signal were received for a blocking
borrow not done on the writer locker's thread and that borrower's thread
were interrupted, all further blocking borrows could deadlock, never
receiving a 'notEmpty' signal even after the write thread were unlocked.
Also, if a 'notEmpty' signal were received for a timed borrow that timed
out before the write lock was released, all further timed borrowers in
the queue at the time that the write lock were released would run the risk
of timing out and failing, even though elements might be available in
the queue, due to never receiving a 'notEmpty' signal.