(SERVER-537) Define stop for jruby service
In order to cleanly shut down the jruby service during a HUP or tk
shutdown, there is certain cleanup that needs to happen. This commit
defined a shutdown function for the jruby service that cleans up the
pools during a shutdown. This is accomplished by passing a reason into
the flush-pool functions.