(SERVER-862) Check for item from borrowItemWithTimeout when time expires
This commit changes the logic in borrowItemWithTimeout slightly to allow
far an item to be borrowed if available just after the total time to
wait has expired.  In previous commits, the method would return null
rather than an item if the total time to wait had just expired but an
item was actually available to be borrowed.

This commit also has a minor optimization which moves the check for an
item to be available for borrowing from a while to a do while, only
checking for the item after the first pass through the code to borrow
rather than unnecessarily checking even before a borrow has been
attempted.