(MAINT) subclass ScriptingContainer to avoid ambiguous signatures
The `callMethod` method of JRuby's `ScriptingContainer` class has a
lot of overloaded signatures, many with overlapping arities.  This
can make it difficult for Clojure to determine the correct version of
the method to call.  Seemingly unrelated modifications to the Clojure
code could unexpectedly trigger changes to which signature would end
up being called at runtime.

This commit creates a very thin subclass of ScriptingContainer, which
basically only exists to provide one delegate method (with a unique
method name and a single signature).  This removes any ambiguity in
terms of which method we are trying to call from Clojure.